pyproject.toml
[project]
name = "llm-validation-poc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.122.0",
    "jwt>=1.4.0",
    "ollama>=0.6.1",
    "openai>=2.7.1",
    "peewee>=3.18.3",
    "pwdlib>=0.3.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
]
